Revert "Feat/signup firebase"
