+++ android/settings.gradle
@@ -20,9 +20,6 @@ pluginManagement {
 plugins {
     id "dev.flutter.flutter-plugin-loader" version "1.0.0"
     id "com.android.application" version "7.3.0" apply false
-    // START: FlutterFire Configuration
-    id "com.google.gms.google-services" version "4.3.15" apply false
-    // END: FlutterFire Configuration
     id "org.jetbrains.kotlin.android" version "1.7.10" apply false
 }
