--- firebase.json
@@ -1,1 +1,0 @@
-{"flutter":{"platforms":{"android":{"default":{"projectId":"mc656-final-project","appId":"1:309461612025:android:bb9469cac5443614843928","fileOutput":"android/app/google-services.json"}},"ios":{"default":{"projectId":"mc656-final-project","appId":"1:309461612025:ios:fa028ab81c79b190843928","uploadDebugSymbols":false,"fileOutput":"ios/Runner/GoogleService-Info.plist"}},"macos":{"default":{"projectId":"mc656-final-project","appId":"1:309461612025:ios:fa028ab81c79b190843928","uploadDebugSymbols":false,"fileOutput":"macos/Runner/GoogleService-Info.plist"}},"dart":{"lib/firebase_options.dart":{"projectId":"mc656-final-project","configurations":{"android":"1:309461612025:android:bb9469cac5443614843928","ios":"1:309461612025:ios:fa028ab81c79b190843928","macos":"1:309461612025:ios:fa028ab81c79b190843928","web":"1:309461612025:web:2c39353c01300088843928","windows":"1:309461612025:web:d3e083aad88147e1843928"}}}}}}
