--- ios/Runner/GoogleService-Info.plist
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>API_KEY</key>
-	<string>AIzaSyCHp6aL-Yum1SDRsNidRaTINcsN623e_98</string>
-	<key>GCM_SENDER_ID</key>
-	<string>309461612025</string>
-	<key>PLIST_VERSION</key>
-	<string>1</string>
-	<key>BUNDLE_ID</key>
-	<string>com.example.mc656finalproject</string>
-	<key>PROJECT_ID</key>
-	<string>mc656-final-project</string>
-	<key>STORAGE_BUCKET</key>
-	<string>mc656-final-project.firebasestorage.app</string>
-	<key>IS_ADS_ENABLED</key>
-	<false></false>
-	<key>IS_ANALYTICS_ENABLED</key>
-	<false></false>
-	<key>IS_APPINVITE_ENABLED</key>
-	<true></true>
-	<key>IS_GCM_ENABLED</key>
-	<true></true>
-	<key>IS_SIGNIN_ENABLED</key>
-	<true></true>
-	<key>GOOGLE_APP_ID</key>
-	<string>1:309461612025:ios:fa028ab81c79b190843928</string>
-</dict>
-</plist>
