--- lib/firebase_options.dart
@@ -1,89 +1,0 @@
-// File generated by FlutterFire CLI.
-// ignore_for_file: type=lint
-import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
-import 'package:flutter/foundation.dart'
-    show defaultTargetPlatform, kIsWeb, TargetPlatform;
-
-/// Default [FirebaseOptions] for use with your Firebase apps.
-///
-/// Example:
-/// ```dart
-/// import 'firebase_options.dart';
-/// // ...
-/// await Firebase.initializeApp(
-///   options: DefaultFirebaseOptions.currentPlatform,
-/// );
-/// ```
-class DefaultFirebaseOptions {
-  static FirebaseOptions get currentPlatform {
-    if (kIsWeb) {
-      return web;
-    }
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.android:
-        return android;
-      case TargetPlatform.iOS:
-        return ios;
-      case TargetPlatform.macOS:
-        return macos;
-      case TargetPlatform.windows:
-        return windows;
-      case TargetPlatform.linux:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for linux - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      default:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions are not supported for this platform.',
-        );
-    }
-  }
-
-  static const FirebaseOptions web = FirebaseOptions(
-    apiKey: 'AIzaSyDW5eK2sduL-uyyAgtTvtBL0bMEO3oBiqA',
-    appId: '1:309461612025:web:2c39353c01300088843928',
-    messagingSenderId: '309461612025',
-    projectId: 'mc656-final-project',
-    authDomain: 'mc656-final-project.firebaseapp.com',
-    storageBucket: 'mc656-final-project.firebasestorage.app',
-    measurementId: 'G-9YM5LDCQ9Z',
-  );
-
-  static const FirebaseOptions android = FirebaseOptions(
-    apiKey: 'AIzaSyBHw_vEF92MZM12NVGLc-OmVx19B4BiVPE',
-    appId: '1:309461612025:android:bb9469cac5443614843928',
-    messagingSenderId: '309461612025',
-    projectId: 'mc656-final-project',
-    storageBucket: 'mc656-final-project.firebasestorage.app',
-  );
-
-  static const FirebaseOptions ios = FirebaseOptions(
-    apiKey: 'AIzaSyCHp6aL-Yum1SDRsNidRaTINcsN623e_98',
-    appId: '1:309461612025:ios:fa028ab81c79b190843928',
-    messagingSenderId: '309461612025',
-    projectId: 'mc656-final-project',
-    storageBucket: 'mc656-final-project.firebasestorage.app',
-    iosBundleId: 'com.example.mc656finalproject',
-  );
-
-  static const FirebaseOptions macos = FirebaseOptions(
-    apiKey: 'AIzaSyCHp6aL-Yum1SDRsNidRaTINcsN623e_98',
-    appId: '1:309461612025:ios:fa028ab81c79b190843928',
-    messagingSenderId: '309461612025',
-    projectId: 'mc656-final-project',
-    storageBucket: 'mc656-final-project.firebasestorage.app',
-    iosBundleId: 'com.example.mc656finalproject',
-  );
-
-  static const FirebaseOptions windows = FirebaseOptions(
-    apiKey: 'AIzaSyDW5eK2sduL-uyyAgtTvtBL0bMEO3oBiqA',
-    appId: '1:309461612025:web:d3e083aad88147e1843928',
-    messagingSenderId: '309461612025',
-    projectId: 'mc656-final-project',
-    authDomain: 'mc656-final-project.firebaseapp.com',
-    storageBucket: 'mc656-final-project.firebasestorage.app',
-    measurementId: 'G-J3GHGF668H',
-  );
-
-}
