+++ lib/main.dart
@@ -1,13 +1,7 @@
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:mc656finalproject/firebase_options.dart';
-import 'package:mc656finalproject/screens/login_screen.dart';
+import 'package:mc656finalproject/screens/LoginScreen.dart';
 
-void main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp(
-    options: DefaultFirebaseOptions.currentPlatform,
-  );
+void main() {
   runApp(const MyApp());
 }
