+++ lib/screens/SignUpScreen.dart
@@ -1,69 +1,29 @@
 import 'package:flutter/material.dart';
-import 'package:mc656finalproject/screens/home_screen.dart';
-import 'package:mc656finalproject/screens/sign_upScreen.dart';
+import 'package:mc656finalproject/screens/HomeScreen.dart';
 import 'package:mc656finalproject/services/login_verify.dart';
-import 'package:mc656finalproject/utils/colors.dart';
 import 'package:mc656finalproject/components/AppTextField.dart';
+import 'package:mc656finalproject/utils/colors.dart';
 
-class LoginScreen extends StatefulWidget {
-  const LoginScreen({super.key});
+class SignUpScreen extends StatefulWidget {
+  const SignUpScreen({super.key});
 
   @override
-  _LoginScreenState createState() => _LoginScreenState();
+  _SignUpScreen createState() => _SignUpScreen();
 }
 
-class _LoginScreenState extends State<LoginScreen> {
+class _SignUpScreen extends State<SignUpScreen> {
   // Controladores para capturar o texto dos campos de usuário e senha
   final TextEditingController _usernameController = TextEditingController();
   final TextEditingController _passwordController = TextEditingController();
+  final TextEditingController _confirmPasswordController = TextEditingController();
 
-  bool isEmpty() {
-    return (_usernameController.text.isEmpty || _passwordController.text.isEmpty);
-  }
-
-  Future<bool> resultadoLogin() async {
-    Map<String, dynamic> resultado = await LoginVerify(_usernameController.text, _passwordController.text);
-    return resultado['success'];
-  }
-
-  Future<String> mensagemLogin() async {
-    Map<String, dynamic> resultado = await LoginVerify(_usernameController.text, _passwordController.text);
-    return resultado['message'];
-  }
-
-  Future<bool> _login() async {
-    // Lógica para autenticação
-    if (isEmpty()) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text("Por favor, preencha todos os campos")),
-      );
-      return false;
-    } 
-    if (!await resultadoLogin()) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text(await mensagemLogin())),
-      );
-      return false;
-    }
-
-    Navigator.pushReplacement(
-      context,
-      MaterialPageRoute(builder: (context) => HomeScreen()),
-    );
+  // Função de login
+  Future<bool> _signUp() async {
+    final username = _usernameController.text;
+    final password = _passwordController.text;
+    final confirmPassword = _confirmPasswordController.text;
 
     return true;
-  }
-
-
-  void _signUp() async {
-        Navigator.push(
-      context,
-      MaterialPageRoute(builder: (context) => SignUpScreen()),
-    );
-  }
-
-  void _forgotPassword() {
-    print("Forgot my password not developed yet");
   }
 
   @override
@@ -95,6 +55,7 @@ class _LoginScreenState extends State<LoginScreen> {
                     width: 250, // Ajuste o tamanho da logo
                     height: 250,
                   ),
+
                   const SizedBox(height: 24.0),
                   // Campo de usuário
                   AppTextField(
@@ -116,51 +77,23 @@ class _LoginScreenState extends State<LoginScreen> {
                     password: true,
                   ),
                   const SizedBox(height: 16.0),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween, 
-                    children: [
-                      // Botão Entrar alinhado à esquerda
-                      ElevatedButton(
-                        onPressed: _login, 
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: darkPink,
-                          foregroundColor: Colors.white,
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(30.0),
-                            side: const BorderSide(color: darkPink, width: 2.0),
-                          ),
-                          padding: const EdgeInsets.symmetric(vertical: 12.0, horizontal: 24.0),
-                        ),
-                        child: const Text(
-                        'Entrar',
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontWeight: FontWeight.bold,
-                        ),
-                    ),
-                      ),
-                      // Texto "Esqueci minha senha" alinhado à direita
-                      GestureDetector(
-                        onTap: _forgotPassword,
-                        child: const Text(
-                          'Esqueci minha senha',
-                          style: TextStyle(
-                            color: lightPink,
-                            decoration: TextDecoration.underline,
-                          ),
-                        ),
-                      ),
-                    ],
+                  AppTextField(
+                    controller: _confirmPasswordController,
+                    text: 'Confirme a senha', 
+                    vPadding: 10.0, 
+                    hPadding: 20.0, 
+                    bRadius: 30.0,
+                    password: true,
                   ),
                   const SizedBox(height: 16.0),
                   // Botão de Cadastro
                   OutlinedButton(
                     onPressed: () {
                       // Ação de cadastro
-                      _signUp();
+                      print("Volta para a tela de login");
                     },
                     style: OutlinedButton.styleFrom(
-                      backgroundColor: Colors.white, 
+                      backgroundColor: Colors.white, // Cor do texto do botão
                       side: const BorderSide(color: darkPink, width: 2.0),
                       shape: RoundedRectangleBorder(
                         borderRadius: BorderRadius.circular(30.0),
