+++ lib/services/signUp.dart
@@ -1,12 +1,9 @@
 import 'dart:convert';
 import 'dart:io';
-import 'package:firebase_auth/firebase_auth.dart';
-
 import 'password_service.dart';
 import 'username_check.dart';
 
 class SignUp {
-  final FirebaseAuth _auth = FirebaseAuth.instance;
   static Future<bool> register(String username, String password) async {
     // Checagem da senha fornecida
     if (!PasswordService.isStrongPassword(password)) {
@@ -34,29 +31,5 @@ class SignUp {
 
     print('Cadastro realizado com sucesso.');
     return true;
-  }
-
-  Future<bool> registerWithEmailPassword(String email, String password) async {
-    try {
-      UserCredential userCredential =
-          await _auth.createUserWithEmailAndPassword(
-        email: email,
-        password: password,
-      );
-      print('Cadastro realizado com sucesso: ${userCredential.user?.uid}');
-      return true;
-    } on FirebaseAuthException catch (e) {
-      if (e.code == 'weak-password') {
-        print('A senha fornecida é muito fraca.');
-      } else if (e.code == 'email-already-in-use') {
-        print('A conta já existe para esse email.');
-      } else {
-        print('Erro ao cadastrar usuário: $e');
-      }
-      return false;
-    } catch (e) {
-      print('Erro inesperado: $e');
-      return false;
-    }
   }
 }
