+++ macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,10 +5,6 @@
 import FlutterMacOS
 import Foundation
 
-import firebase_auth
-import firebase_core
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  FLTFirebaseAuthPlugin.register(with: registry.registrar(forPlugin: "FLTFirebaseAuthPlugin"))
-  FLTFirebaseCorePlugin.register(with: registry.registrar(forPlugin: "FLTFirebaseCorePlugin"))
 }
