+++ pubspec.yaml
@@ -35,8 +35,6 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
-  firebase_core: ^3.4.0
-  firebase_auth: ^5.2.0
 
 dev_dependencies:
   flutter_test:
