+++ windows/flutter/generated_plugins.cmake
@@ -3,8 +3,6 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
-  firebase_auth
-  firebase_core
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
